move to aggregates submodule
