+++ discovery-provider/integration_tests/queries/test_populate_track_metadata.py
@@ -3,7 +3,7 @@ import logging
 from integration_tests.utils import populate_mock_db
 from src.queries import response_name_constants
 from src.queries.query_helpers import populate_track_metadata
-from src.tasks.index_aggregate_track import _update_aggregate_track
+from src.tasks.aggregates.index_aggregate_track import _update_aggregate_track
 from src.utils.db_session import get_db
 
 logger = logging.getLogger(__name__)
