+++ discovery-provider/integration_tests/queries/test_search_track_tags.py
@@ -1,6 +1,6 @@
 from integration_tests.utils import populate_mock_db
 from src.queries.search_track_tags import search_track_tags
-from src.tasks.index_aggregate_plays import _update_aggregate_plays
+from src.tasks.aggregates.index_aggregate_plays import _update_aggregate_plays
 from src.utils.db_session import get_db
