+++ discovery-provider/integration_tests/tasks/test_index_aggregate_plays.py
@@ -4,7 +4,7 @@ from typing import List
 
 from integration_tests.utils import populate_mock_db
 from src.models import AggregatePlays
-from src.tasks.index_aggregate_plays import _update_aggregate_plays
+from src.tasks.aggregates.index_aggregate_plays import _update_aggregate_plays
 from src.utils.db_session import get_db
 
 logger = logging.getLogger(__name__)
