+++ discovery-provider/src/app.py
@@ -368,7 +368,7 @@ def configure_celery(celery, test_config=None):
             "src.tasks.index_plays",
             "src.tasks.index_metrics",
             "src.tasks.index_materialized_views",
-            "src.tasks.index_aggregate_plays",
+            "src.tasks.aggregates.index_aggregate_plays",
             "src.tasks.index_aggregate_monthly_plays",
             "src.tasks.index_hourly_play_counts",
             "src.tasks.vacuum_db",
@@ -380,7 +380,7 @@ def configure_celery(celery, test_config=None):
             "src.tasks.index_solana_plays",
             "src.tasks.index_aggregate_views",
             "src.tasks.index_aggregate_user",
-            "src.tasks.index_aggregate_track",
+            "src.tasks.aggregates.index_aggregate_track",
             "src.tasks.index_challenges",
             "src.tasks.index_user_bank",
             "src.tasks.index_eth",
