+++ discovery-provider/src/tasks/aggregates/index_aggregate_plays.py
