+++ discovery-provider/src/tasks/aggregates/index_aggregate_track.py
